Switch text and blob types
Previously, I was using bytes for text and a subclass of bytes for
blobs.  Switch to using unicode for text and bytes for blobs.  This is
easier to use in almost all cases, with the exception of if you need to
interact with a text column that stores non-UTF-8 data.

 - bind a blob value with "%(x)s" and x=b'foo'
 - bind a UTF-8 string value with "%(x)s" and x=u'foo'
 - bind a non-UTF-8 string with "cast(%(x)s as text)" and x=b'foo'

 - fetch a blob as bytes with "select x"
 - fetch a UTF-8 string as unicode with "select x"
 - fetch a non-UTF-8 string as bytes with "select cast(x as blob)"