Check for closed handles without function calls
Rather than paying a price to check whether the handle has been closed
on each operation, instead have close replace all public methods with
a stub that raises a "not connected" error.  If any of these functions
is called after close was called, an error will be raised automatically.

Note that this can't be applied to __iter__ and __next__, because
special methods are looked up on the class, not the instance, as
a CPython optimization - see
https://docs.python.org/3/reference/datamodel.html#special-method-lookup

So, instead, we simply set _more_rows_available to False, so that any
attempt to iterate over rows will immediately raise a StopIteration
exception.