Swap Connection to weakly reference Cursor instead
Previously Connections would keep Cursors alive, but not vice versa.
Upon reflection, that's probably the opposite of what we want.  It's
still possible to use a Connection through a Cursor (via .connection)
even if you've lost the original reference to the Connection.  The
reverse isn't true - there's nothing useful you can do to a Cursor if
you haven't held onto a reference to it.

Update the test to assert the new behavior, and factor some common code
in Connection into helper functions.