Block naïve datetimes as parameters
Comdb2 doesn't store timezones.  Instead, it takes your input
date+time+timezone and normalizes them to a point along an idealized
timeline.  When you retrieve a datetime, it performs the same
transformation in the reverse direction, to whatever timezone is
requested by your connection.

In Python, a datetime can either be aware (meaning it knows what
timezone it's for), or naïve (meaning that it doesn't).

Our old behavior was to treat naïve datetimes as local time, but this
leads to time appearing to flow backwards around DST switches.  We
wouldn't be able to tell whether a datetime corresponded to the first or
second 2016-11-06T01:30:00; and we'd always treat it as the first.

As aware datetimes can unambiguously interpreted as a point along an
idealized timeline and naïve datetimes can't, this changeset changes our
behavior to raise an exception if a naïve datetime is passed as
a parameter to execute().