Implement `callproc`, in terms of `execute`
This adds the optional DB-API 2.0 `callproc` method.  It simply
constructs a valid "exec procedure" statement corresponding to the
provided sequence of parameters, reformats them as a dictionary of
parameters, and passes them on to `execute`.

Because Comdb2 does not allow binding a procedure name, we also validate
that the procedure name matches a small set of characters known to be
allowed in SP names.

This potentially allows 3rd party code to call Comdb2 stored procedures.
Note, though, that the major impetus for including the `callproc` method
in DB-API appears to have been defining how output parameters are
handled, and Comdb2 does not support output parameters.