Add test for reading/writing datetimes
The previous tests all round tripped the datetime.  This test ensures
that sending a datetime to the DB works independently of retrieving
a datetime from the DB, and vice versa - so that there's no way a bug in
how I create a cdb2_client_datetime_t could be cancelled out by
a matching bug in reading a cdb2_client_datetime_t.