Update a test relying on caught exceptions' scopes
In python 2, a name assigned to an exception in an exception handler
remains bnound even after the handler exits.

In python 3, this is no longer the case, and you need to explicitly
assign a name to the exception if you want it to stick around.  This is
to prevent reference cycles (the exception storing a traceback storing
a stack frame, and the stack frame storing a reference to the
exception).

Get around the change by checking the caught exception's value in the
exception handler, and adding an extra assert to ensure that an
exception was in fact raised.