Handle Python 3 ffi.string returning byte strings
Comdb2 expects and returns all strings as UTF-8 byte strings.  We want
to map those onto the Python `str` type.  For `char *`s returned by
Comdb2, `ffi.string` creates byte strings to hold them.  We're happy
with that for Python 2 (where a byte string is the `str` type), but not
for Python 3 (where a unicode string is the `str` type).

So, handle this by defining a function, `_ffi_string`, that wraps
`ffi.string` but returns its output as a `str` unconditionally.  On
Python 2, this is the same as calling `ffi.string`.  On Python 3, it's
a call to `ffi.string` plus an additional `bytes.decode` call.