Don't check column types for empty result sets
It appears that the column types given by Comdb2 for empty result sets
aren't stable over versions.  Fix two places where we were checking the
column types of an empty result set.  In the first spot, explicitly
check only the column name and not the column type.  In the second spot,
add a COMMIT before the SELECT so that the result set won't be empty.