Ensure lib.cdb2_close is accessible during __del__
Due to unspecified order of finalization for global variables, it was
possible for `Handle.__del__` to fail because it tries to call
`lib.cdb2_close` and the global `lib` variable had already been
finalized.  Work around this by saving a reference to `lib.cdb2_close`
for use in `__del__`.

Also reduce the work done by `Handle.__del__`.  Previously, it was
written in terms of `Handle.close`, but that does more than just close
the underlying connection; it also updates instance methods and
variables to new values, which is pointless when the instance is being
finalized.

Reported-by: DRQS 88746293
Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>
Signed-off-by: Eric N. Vander Weele <evanderweele@bloomberg.net>