Remove `dbapi2.Connection.__del__` finalizer
This finalizer doesn't add any value.  Its only purpose was to ensure
that the low level Comdb2 connection was closed gracefully and could be
donated back to the socket pool.  But, even without this destructor,
everything works properly: first, any `Cursor` that is still open is
garbage collected (before its associated `Connection`, since the
`Cursor` holds a reference to its `Connection`).  That breaks the
reference to the `Connection` that the `Cursor` was holding and allows
the `Connection` to be garbage collected.  That breaks the reference to
the low level `Handle` that the `Connection` was holding and allows the
`Handle` to be garbage collected.  `cdb2.Handle.__del__` takes care of
donating the socket back to sockpool from there.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>
Signed-off-by: Eric N. Vander Weele <evanderweele@bloomberg.net>