Figure out the SQL operation in a single place
Previously we could run two regex searches, one to figure out if the SQL
is for transaction management, and a second one to check if it is
a client-side SET statement or if it talks to the server.  Consolidate
that into a single regex search, and pass the operation along as needed.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>