Add a dict mapping known connect flags to values
Add a dict that maps all known flags for the cdb2_open call to their
values, so that named constants can be passed to the Handle constructor
(or OR'd together, if multiple flags need to be passed).

Also, update Handle's constructor to refer to DIRECT_CPU by name rather
than as a magic number, and add a test that uses flags, using the one
flag that has an effect that we can observe.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>