Allow retrieving rows as dicts or namedtuples
Add a new property, `row_factory` both dbapi2.connect and cdb2.Handle
allowing users to specify a factory for changing the returned row type.
The factory function is constructed with a list of column names, and the
callable,  returned by the factory, is called with a list of column
values for producing the resulting row.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>