Improve _TypeObject.__repr__
Sphinx wants to stringify the type objects for use in the generated
documentation.  By updating _TypeObject.__repr__, we can make them look
nice even when printed in the REPL or by help().

To accomplish this, we'll construct _TypeObject with type names, rather
than type enum values.  We'll hold on to those constructor arguments,
and format them into our repr.

This is a backwards incompatible change to how a private class is
constructed, but no one but us should ever be constructing one of these.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>