Exclude interval types from documentation mock
The entire purpose of this mock is to allow the comdb2.cdb2 module to be
imported by Sphinx's autodoc so that it can introspect documentation.
With that in mind, we should exclude the 3 Comdb2 interval types from
the mock, which we agreed not to support in our RFC - having them adds
no value to the documentation and can only mislead readers.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>