Stop consuming remaining rows when decoding fails
The old behavior was inconsistent with how we handled other types of
errors (for instance, failing to decode a row's columns would consume
all remaining rows, but failing to convert a row using the user
specified row factory wouldn't).  It's also unnecessary: if the user
wants to run another query, they'll be able to - and at that point the
previous result set will be consumed.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>