Test timezones where tzname() isn't in Olson DB
pytz returns either "EST" or "EDT" as the tzname for America/New_York
datetimes, depending on whether they are in or out of daylight savings
time.  Add a test showing that this causes issues when we send such
datetimes to a Comdb2, as it doesn't recognize "EDT" (because it isn't
in the Oslon database).

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>