Invalidate cursors in more places
Previously, an old cursor was only getting invalidated when a new cursor
was created.  Instead, it should be invalidated whenever the user
attempts to execute a new query (whether or not it succeeds), as well as
whenever `get_effects` is called (because it consumes the old result
set).  Trying to consume an old iterator after either of those points is
a usage error, and we should raise an exception indicating so.