Make namedtuple_factory handle DML
Previously, this would fail with a ValueError because Comdb2 would
respond with a column name like "rows inserted", which namedtuple
couldn't use as a field name.

Since the user wasn't in control of that column name, the best solution
is to mangle the column name into something that namedtuple won't choke
on.  We've chosen `_0`, which is what the namedtuple factory function
mangles it to when `rename=True` is given.  We may choose to do
`rename=True` unconditionally in the future, but for now we're limiting
it to this single specific case.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>