Rename package to 'comdb2' from 'python-comdb2'
Even if the name of the repository is 'python-comdb2' there is no need
for the name of the Python package to be 'python-comdb2'. Should this be
distributed into a dpkg or anything outside of an 'only-python-packages'
system that package will probably well named 'python-comdb2'.

This package is not yet published into PyPI and any user who has locally
installed via `python setup.py install` will just see his files
overridden with the new package name. The change should be safe.

Signed-off-by: Mario Corchero <mcorcherojim@bloomberg.net>