Remove use of cython at setup.py
comdb2 has a dependency on cython as part of its `setup_requires` but
`setup.py` is importing cython which makes it useless.

Since there is already a requirement for setup_tools >= 18.0 there is
no need to cythonise the code, as setuptools support cython.
It is actually more correct not to do it so the sdist will contain
the source cython files.

setup_tools works the issue around by deferring the use of cython after
the dependencies are resolved and installed.

This is an issue as users willing to install comdb2 need to have
cython already preinstalled in their system (before even the
requirement of setup_requires is read).

Signed-off-by: Mario Corchero <mcorcherojim@bloomberg.net>