Handle SQL statements beginning with a comment
`comdb2.dbapi2` needs to determine whether a SQL statement is a SET vs
a SELECT vs an INSERT, because they all are required to behave slightly
differently - for a SET, we can't send a BEGIN before sending the
statement, and for an INSERT we can ask for a rowcount but for a SELECT
we can't lest our result set be consumed, etc.  Since we don't include
a client-side SQL parser, up until now we've just been looking at the
first word of the query.

This doesn't handle the case where the query begins with a comment.  Use
a more complicated regex to attempt to skip over comments and grab the
first word that isn't part of a comment, instead.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>