Fix `DatetimeUs.astimezone` Python 3 compatibility
In Python 3, it's legal to call `datetime.datetime.astimezone` with no
arguments, in which case the system's local UTC offset is used for the
returned `datetime.datetime` object. Allow this to work for `DatetimeUs`
as well, by allowing it to be called with any number of arguments, and
forwarding them blindly along to `datetime.datetime` for validation.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>