stubs: Type parameter dict's keys as Text
The `Mapping` type is invariant for the key, rather than covariant,
meaning that a `Mapping[Text, int]` isn't considered a subtype of
`Mapping[Union[Text, bytes], int]`, so these annotations were wrong.

To fix them, we'll say that the key must be `Text`.

For the DB-API 2.0 interface this makes no difference whatsoever,
because in Python 3 mode we already only accepted `Text`, and in Python
2 mode mypy accepts a `bytes` where a `Text` was expected.

For the `cdb2` submodule, this means that using `bytes` for the key in
Python 3 won't typecheck cleanly, even though it works. That seems like
a reasonable tradeoff: although it works, we'd prefer users use `str`
keys, so the warning if they don't seems reasonable.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>