Don't consume events when a handle is GC'd
Python makes no guarantees about whether finalizers are called
for objects that still exist when the main script ends. Because of this,
it's unreasonable to have a `__del__` method that consumes events off of
a Lua consumer queue when it's called. It's not reasonable to have
a function that may or may not be automatically called, but that mutates
the state of a database if it is called.

This is technically a behavior change, but it affects only handles that
were leaked without the user calling `close()` on them. If this changes
the behavior of a user's program, the most likely failure mode is
a crash loop, where the same event is re-delivered to the application
repeatedly without ever being consumed, and each time it is delivered
the application fails to handle it and dies. Any application that fails
in that way was already behaving out of contract, because it was relying
on `__del__` being called.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>