Make consuming events on close() optional
Provide an optional keyword argument on `dbapi2.Connection.close()` and
`cdb2.Handle.close()` that allows the user to control whether the last
event delivered to a Lua consumer will be consumed (when
`consume_current_event=True`, the default) or will be left in the queue
to be redelivered (when `consume_current_event=False`).

This does not change the default behavior. It just allows users of Lua
consumers a way to gracefully `close()` a connection without marking the
last received event as consumed.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>