docs: Fix Sphinx warnings
We were using `T` as a stand-in for any valid parameter type, but Sphinx
doesn't like that. Let's just use `Any` in the docs - the actual valid
types are spelled out in the type stubs, but adding them to the
documentation would make things unnecessarily verbose and repetitive.

Likewise, we were using `Row` as a stand-in for whatever row type is
configured using the `row_factory`, but let's just drop that and leave
it implicit.

Finally, use code formatting (double backticks) rather than references
(single backticks) for some things that were never meant to be
references.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>