Remove `.pxd` source files and fix CI
It shouldn't have been here to begin with, and with newer versions of
Cython this is now an error.

Running CI daily would have helped us detect that issue earlier. This
commit also drops CI for 2.7, 3.6, and 3.7, in preparation for a next
version that will no longer support these versions.

Signed-off-by: Gus Monod <gmonod1@bloomberg.net>