Let users request specific result set column types
Expose a `column_types` keyword-only argument to `cdb2.Handle.execute`
and `dbapi2.Cursor.execute` for specifying the types that the result set
columns must have. The database will coerce returned values to the given
type for the column. An exception will be raised if it can't, or if too
few or too many types are given.

Also, populate an `enum.IntEnum` with the acceptable column types.
Previously we had this only as the members of a dict, but that doesn't
play nicely with type annotations. Keep the dict for backwards
compatibility, but use the enum to populate it.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>