docs: Replace auto-generated NamedTuple docstrings
Without this change, Sphinx complains that these attributes are
documented twice, once by our Sphinx markup in the class docstring and
once with a generated docstring for each property produced by
`NamedTuple` itself for each property.

Override the generated one with our description so that Sphinx finds
only one docstring for each attribute, and so that `help()` shows the
more useful one.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>