Handle C++ exceptions thrown within `libcdb2api`
While `libcdb2api` itself is implemented in C, the version used
internally at Bloomberg can optionally call into some plugins that are
implemented in C++, which could raise exceptions.

Ensure that `python-comdb2` catches any such exceptions, both for the
benefit of Bloomberg's internal Comdb2 users and to guard against the
possibility that C++ code is added to `libcdb2api` itself in the future.

Signed-off-by: Sarah Monod <smonod1@bloomberg.net>