docs: Remove an unused setting from docs/conf.py
This is now triggering warnings saying that:

    WARNING: The config value `viewcode_import' has type `bool';
             expected `NoneType'.

The comment says that we were trying to prevent the `viewcode` extension
from generating HTML for the standard library `datetime` module, but
even with this setting removed it doesn't do that. This may just be
a holdover from some long ago version of Sphinx, but it no longer seems
to be necessary or useful.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>