Fix binding of ambiguous datetimes with `fold=1`
When binding a `datetime.datetime` (but _not_ a `DatetimeUs`), the
`fold` attribute was incorrectly ignored, leading to the first
occurrence of the given wall clock time being bound, rather than the
later occurrence. This was caused by the binding layer rounding the
parameter value to the nearest millisecond by adding a `timedelta` to
it, which resets the `fold` attribute to 0.

Work around this by converting to UTC first, before rounding to the
nearest millisecond. This works since there are no ambiguous times in
UTC and so the `fold` attribute must always be zero.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>