Mark cursor drain test as such
Not fully draining a cursor is usually a programming error, and we want
to avoid it from happening. To that end, alarms have been setup and
caught this particular test as being an offender.

However, in this specific scenario, we do want to leave the cursor
undrained, because this is precisely the behavior that we want to test,
especially where it relates to the lifecycle of the underlying
connection and garbage collection. Therefore, we are marking this test
as a false positive to avoid having the alarm raised for this code.

Signed-off-by: Sarah Monod <smonod1@bloomberg.net>